+++ ignite/cmd/account_create.go
@@ -16,6 +16,8 @@ func NewAccountCreate() *cobra.Command {
 		RunE:  accountCreateHandler,
 	}
 
+	c.Flags().AddFlagSet(flagSetCoinType())
+
 	return c
 }
 
@@ -29,6 +31,7 @@ func accountCreateHandler(cmd *cobra.Command, args []string) error {
 	ca, err := cosmosaccount.New(
 		cosmosaccount.WithKeyringBackend(getKeyringBackend(cmd)),
 		cosmosaccount.WithHome(getKeyringDir(cmd)),
+		cosmosaccount.WithCoinType(getCoinType(cmd)),
 	)
 	if err != nil {
 		return errors.Errorf("unable to create registry: %w", err)