+++ ignite/cmd/account_show.go
@@ -25,6 +25,7 @@ func accountShowHandler(cmd *cobra.Command, args []string) error {
 	ca, err := cosmosaccount.New(
 		cosmosaccount.WithKeyringBackend(getKeyringBackend(cmd)),
 		cosmosaccount.WithHome(getKeyringDir(cmd)),
+		cosmosaccount.WithBech32Prefix(getAddressPrefix(cmd)),
 	)
 	if err != nil {
 		return err