+++ ignite/cmd/testnet_inplace.go
@@ -30,6 +30,8 @@ We can create a testnet from the local network state and mint additional coins f
 	c.Flags().AddFlagSet(flagSetCheckDependencies())
 	c.Flags().AddFlagSet(flagSetSkipProto())
 	c.Flags().AddFlagSet(flagSetVerbose())
+	c.Flags().AddFlagSet(flagSetAccountPrefixes())
+	c.Flags().AddFlagSet(flagSetCoinType())
 
 	c.Flags().Bool(flagQuitOnFail, false, "quit program if the app fails to start")
 	return c
@@ -71,17 +73,27 @@ func testnetInplace(cmd *cobra.Command, session *cliui.Session) error {
 	if err != nil {
 		return err
 	}
+
 	home, err := c.Home()
 	if err != nil {
 		return err
 	}
+
 	keyringBackend, err := c.KeyringBackend()
 	if err != nil {
 		return err
 	}
+
+	prefix := getAddressPrefix(cmd)
+	addressCodec := address.NewBech32Codec(prefix)
+	valAddressCodec := address.NewBech32Codec(prefix + "valoper")
+	coinType := getCoinType(cmd)
+
 	ca, err := cosmosaccount.New(
 		cosmosaccount.WithKeyringBackend(cosmosaccount.KeyringBackend(keyringBackend)),
 		cosmosaccount.WithHome(home),
+		cosmosaccount.WithBech32Prefix(prefix),
+		cosmosaccount.WithCoinType(coinType),
 	)
 	if err != nil {
 		return err
@@ -93,9 +105,6 @@ func testnetInplace(cmd *cobra.Command, session *cliui.Session) error {
 		accErr          *cosmosaccount.AccountDoesNotExistError
 	)
 
-	prefix := getAddressPrefix(cmd)
-	addressCodec := address.NewBech32Codec(prefix)
-	valAddressCodec := address.NewBech32Codec(prefix + "valoper")
 	for _, acc := range cfg.Accounts {
 		sdkAcc, err := ca.GetByName(acc.Name)
 		if errors.As(err, &accErr) {
@@ -138,5 +147,6 @@ func testnetInplace(cmd *cobra.Command, session *cliui.Session) error {
 		NewOperatorAddress: operatorAddressStr,
 		AccountsToFund:     accounts,
 	}
+
 	return c.TestnetInPlace(cmd.Context(), args)
 }