+++ ignite/templates/app/app.go
@@ -38,6 +38,7 @@ func NewGenerator(opts *Options) (*genny.Generator, error) {
 		excludePrefix []string
 		overridesFS   = make(map[string]embed.FS)
 	)
+
 	if opts.IsChainMinimal {
 		// minimal chain does not have ibc
 		excludePrefix = append(excludePrefix, ibcConfig)
@@ -68,6 +69,7 @@ func NewGenerator(opts *Options) (*genny.Generator, error) {
 	ctx.Set("GitHubPath", opts.GitHubPath)
 	ctx.Set("BinaryNamePrefix", opts.BinaryNamePrefix)
 	ctx.Set("AddressPrefix", opts.AddressPrefix)
+	ctx.Set("CoinType", opts.CoinType)
 	ctx.Set("DepTools", cosmosgen.DepTools())
 	ctx.Set("IsChainMinimal", opts.IsChainMinimal)